start impl. get appliances enum
