+++ Assignment1/helper_functions.py
@@ -1,13 +1,23 @@
-import numpy as np
+import pandas as pd
+import enum
+from classes import Appliance
 
-def get_pricing(pricing: str) -> np.ndarray:
-    if pricing == "ToU":
-        pricing = np.zeros(24)
-        pricing[0:17] = 0.5
-        pricing[17:20] = 1.0
-        pricing[20:24] = 0.5
-    elif pricing == "RTP":
-        pricing = np.load('data/rt_pricing.npy')
-    else:
-        raise ValueError("Pricing must be either 'ToU' or 'RTP'.")
-    return pricing
+def get_appliances_enum() -> enum.Enum:
+    appliances_dict = {}
+
+    df_appliances = pd.read_excel('data/energy_usage.xlsx')
+    for i, row in df_appliances.iterrows():
+        appliance = Appliance(row['Appliances'].replace(" ", ""),
+                              row['Shiftable'],
+                            #   row['Length [h]'],
+                              row['Daily usage [kWh]'],
+                              row['Length [h]'],
+                              row['Alpha'],
+                              row['Beta']
+        )
+        appliances_dict[appliance.name] = appliance
+    return enum.Enum('Appliances', appliances_dict)
+
+if __name__ == "__main__":
+    print(get_appliances_enum())
+    print(get_appliances_enum().Laundrymachine)
